requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
motor
scikit-learn
